+++ R/simpA.NP.R
@@ -172,13 +172,11 @@ simpA.NP <- function(
 
   if (testStat %in% c("T1_CvM_Cs3", "T1_CvM_Cs4", "tilde_T0_CvM",
                       "T1_KS_Cs3", "T1_KS_Cs4", "tilde_T0_KS")){
-    nGrid = numericalInt$nGrid
-    grid <- statmod::gauss.quad(n = nGrid, kind = numericalInt$kind)
-    # Change of range to be on [truncVal , 1 - truncVal]
-    grid$nodes <- grid$nodes * (1/2 - truncVal) + 1/2
 
-    # FIXME: the weights should depend on `truncVal`
-    grid$weights <- grid$weights / 2
+    nGrid = numericalInt$nGrid
+    # grid on [truncVal , 1 - truncVal]
+    grid <- get.gauss.quad(nGrid = nGrid, kind = numericalInt$kind,
+                           center = 1/2, half_length = (1/2 - truncVal))
   }
 
   if (testStat %in% c("T1_CvM_Cs3", "T1_CvM_Cs4", "tilde_T0_CvM",